// src/app/models/seguimiento.ts
export interface Seguimiento {
  id?: string;
  numeroRequerimiento?: number;
  numeroSeguimiento?: number;
  tipoRequerimiento?: string;
  responsable?: string;
  responsableSeguimiento?: string;
  fechaSeguimiento?: string;
  asuntoSeguimiento?: string;
  docAdjunto?: string;
  nombre?: string;
  observaciones?: string;
  estado?: string;
}